+++ src/lib/utils.ts
@@ -1,4 +1,3 @@
-import type { Updater } from "@tanstack/vue-table"
 import type { ClassValue } from "clsx"
 import type { Ref } from "vue"
 import { clsx } from "clsx"
@@ -8,9 +7,8 @@ export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
 
-export function valueUpdater<T extends Updater<any>>(updaterOrValue: T, ref: Ref) {
-  ref.value
-    = typeof updaterOrValue === "function"
-      ? updaterOrValue(ref.value)
-      : updaterOrValue
+export function valueUpdater<T>(updaterOrValue: T | ((old: T) => T), ref: Ref) {
+  ref.value = typeof updaterOrValue === "function"
+    ? (updaterOrValue as (old: T) => T)(ref.value)
+    : updaterOrValue
 }